Remove Firebase admin credentials and initialization
Removed Firebase admin credentials and initialization code, as it poses
a security risk to have sensitive information stored in the codebase.
This change enhances the security posture of the application by
following best practices for handling sensitive data. This also prevents
exposure of confidential information in the event of a code leak. Issue
#123, #124